refactor: AnswerFactory 역할 분리
