--- src/main/java/sehee/util/NumberMaker.java
@@ -1,13 +1,0 @@
-package sehee.util;
-
-import java.util.Random;
-
-public record NumberMaker(
-    Random random
-) {
-
-    public int makeRandomOne() {
-        return random.nextInt(9) + 1;
-    }
-
-}
